修复了delete的重大问题（大概）

+++ src/com/xiaohuo/libmanager/services/BooksManageService.java
@@ -8,6 +8,7 @@ import com.xiaohuo.libmanager.services.template.Journal;
 import com.xiaohuo.libmanager.services.template.Newspaper;
 import com.xiaohuo.libmanager.services.template.TypeList;
 
+import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -307,40 +308,35 @@ public class BooksManageService
      * Delete book in the database according to their title
      * @param title the title of the book the user want to delete
      */
-    public void deleteBookByTitle(String title) throws CollectionException{
+    public void deleteBook(String title) throws CollectionException{
         BooksManageDao dao = new BooksManageDao();
+        Map<Integer,ArrayList<String>>bookList;
         List<Throwable> exceptions = new ArrayList<>();
-        try
-        {
-            dao.delete(title);
+        bookList = dao.search(title);
+        if(bookList.size()!=0){
+            ArrayList<String> book;
+            if(bookList.size()==1){
+                book = bookList.get(0);
+            }
+            else {
+                // ask use to choose one book
+                int key;
+                book = bookList.get(key);
+            }
+            String code = book.get(4);
+            String type = book.get(3);
+            int ID = getBookID(type,code);
+            try {
+                dao.delete(ID);
+            }
+            catch (CollectionException e){
+                exceptions.add(e);
+            }
         }
-        catch (CollectionException e)
-        {
-            exceptions.add(e);
+        else {
+            System.out.println("Book not found");
         }
-        if(exceptions.size()>0)
-        {
-            throw new CollectionException(exceptions);
-        }
-    }
-
-    /**
-     * Delete book in the database according to its ID
-     * @param ID the title of the book the user want to delete
-     */
-    public void deleteBookByID(String ID) throws CollectionException{
-        BooksManageDao dao = new BooksManageDao();
-        List<Throwable> exceptions = new ArrayList<>();
-        try
-        {
-            dao.delete(ID);
-        }
-        catch (CollectionException e)
-        {
-            exceptions.add(e);
-        }
-        if(exceptions.size()>0)
-        {
+        if(exceptions.size()>0){
             throw new CollectionException(exceptions);
         }
     }
